Initial commit of adjust_screen_brightness python application
Since i3 doesn't enable things like adjusting screen brightness from the
keyboard fn keys, this script is being written to hopefully fix this.  We'll
make i3 call this script when the screen brightness keys are pressed.
It will pass in the pressed keycode and adjust the brightness either up
or down.  This is the initial commit, so nothing working at the moment.
First time playing with python, so it's a bit slow going but getting
there.