Added updated home conky configuration
I've had a bit more lucky figuring out conky, so here is an updated
configuration for conky that I'm using at home.  Need to look at the one
for work a bit more, and also possibly add a second conky instance to
display a few things.  This is a start on figuring everything out.